api/requirements.txt
Flask==2.2.2
flask_cors==3.0.10
plaid-python==13.0.0
